modified:   backend/main.py 	modified:   backend/models/user.py 	new file:   backend/requests.py
